# docs/api/openapi-v1.designed.json
{
  "components": {
    "schemas": {
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "error",
          "request_id"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "JobResponse": {
        "additionalProperties": false,
        "properties": {
          "document_id": {
            "format": "uuid",
            "title": "Document Id",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ]
          },
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "progress": {
            "anyOf": [
              {
                "maximum": 1,
                "minimum": 0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          },
          "state": {
            "enum": [
              "queued",
              "fetching",
              "parsing",
              "chunking",
              "embedding",
              "indexing",
              "ready",
              "failed",
              "cancelled"
            ],
            "title": "State",
            "type": "string"
          },
          "version_id": {
            "format": "uuid",
            "title": "Version Id",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "job_id",
          "document_id",
          "version_id",
          "state",
          "progress",
          "error",
          "retryable"
        ],
        "title": "JobResponse",
        "type": "object"
      },
...
      "SSEEvent": {
        "discriminator": {
          "mapping": {
            "answer_delta": "#/components/schemas/AnswerDeltaEvent",
            "done": "#/components/schemas/DoneEvent",
            "error": "#/components/schemas/ErrorEvent",
            "evidence": "#/components/schemas/EvidenceEvent",
            "meta": "#/components/schemas/MetaEvent"
          },
          "propertyName": "event"
        },
        "oneOf": [
          {
            "$ref": "#/components/schemas/MetaEvent"
          },
          {
            "$ref": "#/components/schemas/EvidenceEvent"
          },
          {
            "$ref": "#/components/schemas/AnswerDeltaEvent"
          },
          {
            "$ref": "#/components/schemas/DoneEvent"
          },
          {
            "$ref": "#/components/schemas/ErrorEvent"
          }
        ]
      },
...
    },
    "securitySchemes": {
      "AppServiceKey": {
        "in": "header",
        "name": "X-RAG-Service-Key",
        "type": "apiKey"
      },
      "UserJWT": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API v1 contract inventory. Public routes mounted in T26; business services require trusted configuration and authentication.",
    "title": "RAG Core API v1 — designed contract",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/jobs/{job_id}": {
      "get": {
        "description": "Mounted in T26; trusted runtime configuration is required. Authenticated app+subject and active session links/ready versions constrain access. Schemas do not grant ownership or readiness.",
        "operationId": "get_job",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Served response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid or missing credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Insufficient role"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not in current authorized scope"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Scope/readiness/idempotency conflict"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Deleted session for its owner"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid input"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Rate limit or bounded queue full",
            "headers": {
              "Retry-After": {
                "description": "Backoff delay in seconds",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Provider error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Dependency unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Timeout"
          }
        },
        "security": [
          {
            "AppServiceKey": [],
            "UserJWT": []
          }
        ],
        "summary": "get job",
        "tags": [
          "public"
        ],
        "x-implementation-status": "IMPLEMENTED",
        "x-implementation-task": "T12/T19",
        "x-served": true
      }
    },
    "/v1/jobs/{job_id}/retry": {
      "post": {
        "description": "Mounted in T26; trusted runtime configuration is required. Authenticated app+subject and active session links/ready versions constrain access. Schemas do not grant ownership or readiness.",
        "operationId": "retry_job",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Served response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid or missing credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Insufficient role"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not in current authorized scope"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Scope/readiness/idempotency conflict"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Deleted session for its owner"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid input"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Rate limit or bounded queue full",
            "headers": {
              "Retry-After": {
                "description": "Backoff delay in seconds",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Provider error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Dependency unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Timeout"
          }
        },
        "security": [
          {
            "AppServiceKey": [],
            "UserJWT": []
          }
        ],
        "summary": "retry job",
        "tags": [
          "public"
        ],
        "x-implementation-status": "IMPLEMENTED",
        "x-implementation-task": "T12/T19",
        "x-served": true
      }
    },
    "/v1/query": {
      "post": {
        "description": "Mounted in T26; trusted runtime configuration is required. Authenticated app+subject and active session links/ready versions constrain access. Schemas do not grant ownership or readiness.",
        "operationId": "query",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            },
            "description": "Served response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid or missing credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Insufficient role"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not in current authorized scope"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Scope/readiness/idempotency conflict"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Deleted session for its owner"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid input"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Rate limit or bounded queue full",
            "headers": {
              "Retry-After": {
                "description": "Backoff delay in seconds",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Provider error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Dependency unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Timeout"
          }
        },
        "security": [
          {
            "AppServiceKey": [],
            "UserJWT": []
          }
        ],
        "summary": "query",
        "tags": [
          "public"
        ],
        "x-implementation-status": "IMPLEMENTED",
        "x-implementation-task": "T24/T26",
        "x-served": true
      }
    },
    "/v1/query/stream": {
      "post": {
        "description": "Mounted in T26; trusted runtime configuration is required. Authenticated app+subject and active session links/ready versions constrain access. Schemas do not grant ownership or readiness.",
        "operationId": "query_stream",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/SSEEvent"
                }
              }
            },
            "description": "Served response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid or missing credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Insufficient role"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not in current authorized scope"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Scope/readiness/idempotency conflict"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Deleted session for its owner"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid input"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Rate limit or bounded queue full",
            "headers": {
              "Retry-After": {
                "description": "Backoff delay in seconds",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Provider error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Dependency unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Timeout"
          }
        },
        "security": [
          {
            "AppServiceKey": [],
            "UserJWT": []
          }
        ],
        "summary": "query stream",
        "tags": [
          "public"
        ],
        "x-implementation-status": "IMPLEMENTED",
        "x-implementation-task": "T25/T26",
        "x-served": true,
        "x-sse-contract": {
          "deltas": "provisional",
          "done": "validated final QueryResponse",
          "early-error": "meta -> error before evidence is allowed",
          "event-ids": "increasing per request",
          "heartbeat": "SSE comment; no event ID",
          "order": "meta -> evidence -> answer_delta* -> done|error",
          "replay": false,
          "runtime": "IMPLEMENTED; scope checks before evidence/delta/done; cancel upstream",
          "wire": "SSE UTF-8 frames: id/event/data lines, blank-line delimiter; data line contains JSON. SSEEvent schema describes a parsed event, not a JSON HTTP response body."
        }
      }
    },
    "/v1/sessions": {
      "post": {
        "description": "Mounted in T26; trusted runtime configuration is required. Authenticated app+subject and active session links/ready versions constrain access. Schemas do not grant ownership or readiness.",
        "operationId": "create_session",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Served response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid or missing credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Insufficient role"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not in current authorized scope"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Scope/readiness/idempotency conflict"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Deleted session for its owner"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid input"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Rate limit or bounded queue full",
            "headers": {
              "Retry-After": {
                "description": "Backoff delay in seconds",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Provider error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Dependency unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Timeout"
          }
        },
        "security": [
          {
            "AppServiceKey": [],
            "UserJWT": []
          }
        ],
        "summary": "create session",
        "tags": [
          "public"
        ],
        "x-implementation-status": "IMPLEMENTED",
        "x-implementation-task": "T10",
        "x-served": true
      }
    },
...
  },
...
}